requirements.txt
# c:\Users\rizky\OneDrive\Dokumen\GitHub\ETL\requirements.txt
pandas
numpy
gspread
google-auth
google-auth-oauthlib
google-auth-httplib2
SQLAlchemy
psycopg2-binary 
pyarrow                           
google-api-python-client